quiz activity intent에서 user id 값 불러오기

+++ app/src/main/java/org/sopt/sopkathon_31th/data/remote/api/quiz/QuizService.kt
@@ -9,6 +9,6 @@ interface QuizService {
     // 개인별 퀴즈 문제 리스트 조회 API
     @GET("/page/quiz/{userId}")
     fun getQuizList(
-        @Path("userId") userId: Long
+        @Path("userId") userId: Int
     ): Call<ResponseQuizDto>
 }
